bcachefs: bch_extent_rebalance.wont_recompress_smaller XXX
Add a new bch_extent_rebalance field to flag data that's compressed, but
won't get smaller with a different compression type.

This fixes a (main?) cause of rebalance spinning/not making progress.

XXX: is this actually needed? perusing the write path it seems we don't
check if the data got smaller than it was when previously compressed

Signed-off-by: Kent Overstreet <kent.overstreet@linux.dev>